Move TestUtilMask
